requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.34
pydantic==2.9.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
firebase-admin==6.5.0
prometheus-fastapi-instrumentator==6.1.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
